+++ src/Business.java
@@ -1,6 +1,6 @@
 public class Business extends Account{
 
-    public Business(String name, String address, String email, int age, float balance) {
+    public Business(String name, String address, String email, int age, double balance) {
         super(name, address, email, age, balance);
         this.sortCode = "24-65-27";
     }
