+++ src/Current.java
@@ -1,6 +1,6 @@
 public class Current extends Account{
 
-    public Current(String name, String address, String email, int age, float balance) {
+    public Current(String name, String address, String email, int age, double balance) {
         super(name, address, email, age, balance);
         this.sortCode = "24-65-32";
     }
