+++ src/ISA.java
@@ -1,6 +1,6 @@
 public class ISA extends Account{
 
-    public ISA(String name, String address, String email, int age, float balance) {
+    public ISA(String name, String address, String email, int age, double balance) {
         super(name, address, email, age, balance);
         this.sortCode = "24-65-69";
         this.interest=0.02f;
